kimi-asr-service/requirements.txt
kimia_infer
pyannote.audio>=3.1.0
torch>=2.1.0
torchaudio>=2.1.0
soundfile>=0.12.1
librosa>=0.10.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
gunicorn>=22.0.0
python-multipart>=0.0.9
